+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2024-11-02
+
+### Changes
+
+---
+
+
+#### `waved_audio_player` - `v1.2.0`
+
+ - Adding timing text
+ - Fixing play button icon
+ - Using custom Error
+   
+---
+
 ## 2024-11-01
 
 ### Changes